Consolidate benchmark tests
Combine benchmark correctness and tuning config coverage in one test module.


Copilot-Session: 261c58af-1f7f-42cb-a4bf-3e878b5890f3
